+++ src/ploneintranet/docconv/client/handlers.py
@@ -1,37 +1,34 @@
 import logging
-from zope.annotation import IAnnotations
+import transaction
 
 from ploneintranet.attachments.attachments import IAttachmentStoragable
 from ploneintranet.attachments.utils import IAttachmentStorage
-from ploneintranet.docconv.client.async import queueDelayedConversionJob
-from ploneintranet.docconv.client.config import PDF_VERSION_KEY
-from ploneintranet.docconv.client.config import PREVIEW_IMAGES_KEY
-from ploneintranet.docconv.client.config import PREVIEW_MESSAGE_KEY
-from ploneintranet.docconv.client.config import THUMBNAIL_KEY
-from ploneintranet.docconv.client.exceptions import ConfigError
-from ploneintranet.docconv.client.fetcher import fetchPreviews
-from ploneintranet.docconv.client.interfaces import IDocconv
+from ploneintranet import api as pi_api
+
+from ploneintranet.async.celeryconfig import ASYNC_ENABLED
+from ploneintranet.async.tasks import GeneratePreview
 
 log = logging.getLogger(__name__)
 
 
-def _update_preview_images(obj, event):
-    annotations = IAnnotations(obj)
-    if PREVIEW_IMAGES_KEY in annotations:
-        del annotations[PREVIEW_IMAGES_KEY]
-    if THUMBNAIL_KEY in annotations:
-        del annotations[THUMBNAIL_KEY]
-    if PDF_VERSION_KEY in annotations:
-        del annotations[PDF_VERSION_KEY]
-    if PREVIEW_MESSAGE_KEY in annotations:
-        del annotations[PREVIEW_MESSAGE_KEY]
-    success = queueDelayedConversionJob(obj, obj.REQUEST)
-    if not success:
-        try:
-            fetchPreviews(obj)
-        except ConfigError as e:
-            log.error('ConfigError: %s' % e)
-    generate_attachment_preview_images(obj)
+def generate_previews_async(obj, event=None):
+    """ Generates the previews by dispatching them to the async service
+    """
+    # Need a commit to make sure the content is there
+    transaction.commit()
+    if ASYNC_ENABLED:
+        generator = GeneratePreview(obj, obj.REQUEST)
+        generator()
+    else:
+        pi_api.previews.generate_previews(obj)
+
+
+def handle_file_creation(obj, event=None):
+    """ Need own subscriber as cdv insists on checking for its
+        custom layout. Also we want our own async mechanism.
+    """
+    generate_previews_async(obj)
+    # pi_api.previews.generate_previews(obj)
 
 
 def generate_attachment_preview_images(obj):
@@ -39,19 +36,22 @@ def generate_attachment_preview_images(obj):
         return
     attachment_storage = IAttachmentStorage(obj)
     for att_id in attachment_storage.keys():
-        docconv = IDocconv(attachment_storage.get(att_id))
-        if not docconv.has_thumbs():
-            docconv.generate_all()
+        attachment = attachment_storage.get(att_id)
+        if not pi_api.previews.has_previews(attachment):
+            generate_previews_async(attachment)
+            # pi_api.previews.generate_previews(attachment)
 
 
-def archetype_added_in_workspace(obj, event):
-    _update_preview_images(obj, event)
+def content_added_in_workspace(obj, event):
+    generate_previews_async(obj)
+    # pi_api.previews.generate_previews(obj)
 
 
-def archetype_edited_in_workspace(obj, event):
+def content_edited_in_workspace(obj, event):
     if obj.REQUEST.form.get('file') or obj.REQUEST.get('method') == 'PUT':
-        _update_preview_images(obj, event)
+        generate_previews_async(obj)
+        # pi_api.previews.generate_previews(obj)
 
 
-def attachmentstoragable_added(obj, event):
-    generate_attachment_preview_images(obj)
+# def attachmentstoragable_added(obj, event):
+#     generate_attachment_preview_images(obj)